Refactor data test helpers
Rearrange the helper functions in the various data test helpers so
each helper relates to a specific concept (tutorial, course, etc.)
and remove a redundant helper which had two implementations.
