+++ DigitalLearningSolutions.Data.Tests/Helpers/SectionContentTestHelper.cs
@@ -21,35 +21,5 @@
                 new { sectionId, sectionNumber }
             );
         }
-
-        public void UpdateDiagnosticAttempts(int tutorialId, int progressId, int diagAttempts)
-        {
-            connection.Execute(
-                @"UPDATE aspProgress
-                        SET DiagAttempts = @diagAttempts
-                        WHERE TutorialID = @tutorialId AND ProgressID = @progressId",
-                new { tutorialId, progressId, diagAttempts }
-            );
-        }
-
-        public void UpdateDiagnosticScore(int tutorialId, int progressId, int score)
-        {
-            connection.Execute(
-                @"UPDATE aspProgress
-                        SET DiagLast = @score
-                        WHERE TutorialID = @tutorialId AND ProgressID = @progressId",
-                new { tutorialId, progressId, score }
-            );
-        }
-
-        public void UpdateDiagnosticStatus(int tutorialId, int customisationId, int status)
-        {
-            connection.Execute(
-                @"UPDATE CustomisationTutorials
-                        SET DiagStatus = @status
-                        WHERE TutorialID = @tutorialId AND CustomisationID = @customisationId",
-                new { tutorialId, customisationId, status }
-            );
-        }
     }
 }
