+++ DigitalLearningSolutions.Data.Tests/Helpers/TutorialContentTestHelper.cs
@@ -45,6 +45,36 @@
             );
         }
 
+        public void UpdateDiagnosticAttempts(int tutorialId, int progressId, int diagAttempts)
+        {
+            connection.Execute(
+                @"UPDATE aspProgress
+                        SET DiagAttempts = @diagAttempts
+                        WHERE TutorialID = @tutorialId AND ProgressID = @progressId",
+                new { tutorialId, progressId, diagAttempts }
+            );
+        }
+
+        public void UpdateDiagnosticScore(int tutorialId, int progressId, int score)
+        {
+            connection.Execute(
+                @"UPDATE aspProgress
+                        SET DiagLast = @score
+                        WHERE TutorialID = @tutorialId AND ProgressID = @progressId",
+                new { tutorialId, progressId, score }
+            );
+        }
+        public void UpdateDiagAssessOutOf(int tutorialId, int diagAssessOutOf)
+        {
+            connection.Execute(
+                @"UPDATE Tutorials
+                     SET DiagAssessOutOf = @diagAssessOutOf
+
+                   WHERE Tutorials.TutorialID = @tutorialId;",
+                new { tutorialId, diagAssessOutOf }
+            );
+        }
+
         public void RemoveCustomisationTutorial(int customisationId, int tutorialId)
         {
             connection.Execute(
@@ -61,6 +91,15 @@
             connection.Execute(
                 @"UPDATE CustomisationTutorials
                         SET Status = @status
+                        WHERE TutorialID = @tutorialId AND CustomisationID = @customisationId",
+                new { tutorialId, customisationId, status }
+            );
+        }
+        public void UpdateDiagnosticStatus(int tutorialId, int customisationId, int status)
+        {
+            connection.Execute(
+                @"UPDATE CustomisationTutorials
+                        SET DiagStatus = @status
                         WHERE TutorialID = @tutorialId AND CustomisationID = @customisationId",
                 new { tutorialId, customisationId, status }
             );
